feat: refactor toast notification and loading state functions into utils.js for better modularity
