+++ public/js/user-form.js
@@ -1,3 +1,5 @@
+import { showToast } from "./utils.js";
+
 document
   .querySelector("form[action='/api/users']")
   .addEventListener("submit", async (e) => {
